DEV-2190 Fix calculation of XML filename
The idea is, given an essence filename, to calculate the basename
of its sidecar XML counterpart.

The method `set_extension` appends the extension or replaces the extension
if there is one. This method is used on the stem of the essence filename.
The idea is to append the ".xml" extension to that stem in order to create
sidecar basename. However, if that stem contains a dot, it will be seen
as an extension and it will be replaced by ".xml".
Example: "/path/to/file.1.mxf" (absolute path): "file.1" (stem) -> "file.xml".

Fix the calculation so that it always appends the ".xml" extension.

Also:
- Add some tests for that calculation.
- Remove some unused imports.